new file:    .env 	modified:   JWT/.env 	modified:   JWT/cmd/users/.env 	modified:   JWT/cmd/users/main.go 	new file:   docker-compose.integration.yml 	modified:   docker-compose.yml 	modified:   forum/article/cmd/articles/.env 	modified:   forum/article/internal/delivery/gin/handlers/article_handler.go 	modified:   forum/article/internal/delivery/gin/handlers/comment_handler.go 	modified:   forum/article/internal/repository/comment_repository.go 	modified:   forum/article/internal/usecase/article_uc.go 	modified:   forum/article/internal/usecase/comment_uc.go 	modified:   frontend/package.json 	modified:   frontend/src/App.js 	modified:   gateway/internal/delivery/gin/router.go 	new file:   integration-tests/Dockerfile 	new file:   integration-tests/go.mod 	new file:   integration-tests/go.sum 	new file:   integration-tests/main_test.go 	new file:   integration-tests/wait-for-mysql.sh 	new file:   mysql/Dockerfile 	new file:   mysql/init.sql 	new file:   realtime-chat/Dockerfile

+++ forum/article/internal/usecase/comment_uc.go
@@ -4,12 +4,16 @@ import (
 	"article/internal/entity"
 	"article/internal/repository"
 	"context"
+	"fmt"
+	"time"
+
+	"github.com/google/uuid"
 )
 
 type CommentUseCase interface {
-	CreateComment(ctx context.Context, comment *entity.Comment) error
+	CreateComment(ctx context.Context, comment *entity.Comment, userID string, articleID, username string) error
 	GetCommentById(ctx context.Context, id string) (*entity.Comment, error)
-	UpdateComment(ctx context.Context, comment *entity.Comment) error
+	UpdateComment(ctx context.Context, comment *entity.Comment, commentText string) error
 	DeleteComment(ctx context.Context, id string) error
 	GetAllComments(ctx context.Context) ([]*entity.Comment, error)
 	GetCommentsByArticleID(ctx context.Context, id string) ([]*entity.Comment, error)
@@ -25,12 +29,22 @@ func (c *commentUseCase) GetCommentsByArticleID(ctx context.Context, articleID s
 }
 
 // CreateComment implements CommentUseCase.
-func (c *commentUseCase) CreateComment(ctx context.Context, comment *entity.Comment) error {
+func (c *commentUseCase) CreateComment(ctx context.Context, comment *entity.Comment, userID string, articleID, username string) error {
+
+	comment.CommentAuthorID = userID
+	comment.CommentID = uuid.New().String()
+	comment.CreatedCommentAt = time.Now()
+	comment.UpdatedCommentAt = time.Now()
+	comment.ArticleID = articleID
+	comment.CommentAuthorName = username
 	return c.commentRepo.CreateComment(ctx, comment)
 }
 
 // DeleteComment implements CommentUseCase.
 func (c *commentUseCase) DeleteComment(ctx context.Context, id string) error {
+	if id == "" {
+		return fmt.Errorf("Invalid comment id")
+	}
 	return c.commentRepo.DeleteComment(ctx, id)
 }
 
@@ -41,11 +55,18 @@ func (c *commentUseCase) GetAllComments(ctx context.Context) ([]*entity.Comment,
 
 // GetCommentById implements CommentUseCase.
 func (c *commentUseCase) GetCommentById(ctx context.Context, id string) (*entity.Comment, error) {
+	if id == "" {
+		return nil, fmt.Errorf("Invalid comment id")
+	}
 	return c.commentRepo.GetCommentById(ctx, id)
 }
 
 // UpdateComment implements CommentUseCase.
-func (c *commentUseCase) UpdateComment(ctx context.Context, comment *entity.Comment) error {
+func (c *commentUseCase) UpdateComment(ctx context.Context, comment *entity.Comment, commentText string) error {
+
+	comment.CommentText = commentText
+	comment.UpdatedCommentAt = time.Now()
+
 	return c.commentRepo.UpdateComment(ctx, comment)
 }
